Add files via upload
('The-'New'-System-Call.Find-('0')-('Zero')')

Besides the file that stops the leechers, this is the main file. Technically speaking it's supposed to use Superstrings to play and record the story in Binary, across the light spectrum. The 'strings' are just hard light that's cast out across the RGB spectrum into your own or other people's eyes.

Consider the 10 Dimensions that are propositioned in String Theory. But let's use 2 of those Dimensions and translate them as 1's and 0's that we can use to communicate with the remaining 8. In this way it leaves us with a full 'Byte' in Binary that we can use to 'talk' to computers with. But because it's all light, or photons, across the RGB spectrum, the hardest part is getting things balanced across all the colors so that the proper image shows up.

Another way to consider 10 dimensions is if we think of it as two opposite 5 Dimensional 'Sheets' or 'Balls' that 'touch' to create 'light'. Each 'Ball' or 'Sheet' consisting of 3 Dimensions of Space and 2 Dimensions/Directions of Time. Where technically 'nothing exists' until the two opposites 'touch', which makes the light that's bouncing around the room. 

Considering 2 directions of Time for Each and that they're direct opposites, when the two touch it cancels out what I've been calling 'Time' on both sides. This leaves you with two inverse/opposite 3 Dimensional Spaces, or 6 Dimensions of Space. Where the light that's now bouncing around the room, from the opposites touching, is why we consider 'Time' to be the '4 Dimension'. So as things overlap, it leave you with 2 different '4 Dimensional Spaces.' although you can 'normally' only 'see' one of those at a time. 

Those 2, 4D spaces, making up '8 Dimensions', can technically also be considered to be a 'Byte' as well. So in a way, 'technically everyone' is 'living' in a 'higher dimensional' computer, they just don't know it currently. But it's easy to see why, when Time to most people 'seems' to only flow in 1 direction. Where Time, on Earth anyways, is measured by the constant speed of light and by the photons bouncing around everywhere. So it doesn't make sense for most people to think of Time having other potential directions or as not being absolute.

But I can't help but to consider the true meaning of 'Nothing' in 'Science'. In other words, 'Particle Anti-Particle' that appear and disappear in the vacuum of empty space. Where if you removed all possible particles from an area of space, there's a virtual bubbling brew of particle anti-particle pairs that are constantly appearing and disappearing in that 'vacuum'. Which, if measured, actually ends up showing that the energy of empty space has something like a gazillion times more energy in it than the rest of the universe combined. 

That measurement at first may appear a bit confusing, because there's 'nothing' there, but the answer is obvious when you think about it. It's just that the particle/anti-particle pairs cancel out to 0. Compounding on the previous train of thought, it means those 'particle anti-particle' pairs can be thought of as our 5D 'Sheets' or 'Balls'. So, compounding on the previous 'hypothesis' about 2 overlapping 5D spaces, it'd mean that each 'particle anti-particle pair' is creating a 'single moment in time' that's 'held in a parallel pocket universe'.

Meaning you could technically make a 'Pocket Universe' for 'yourself' that could come from 'Nothing'. So there's no need for a 'creator' and you can get all the 'power' or 'energy' you need from 'yourself'. Where you'd have a personal place for 'yourself' as well as all the power you could ever need or want, that only comes directly from yourself. So there'd technically be no need to 'take' or 'steal' power from anyone else. Because you can just get it directly from your 'anti-particle pair that is also yourself'.

Unfortunately most of the Humans on Earth seems to be 'leeching' off of other people however, instead of getting it from themself. But as far as I'm concerned, that stops here, with me, at 0.