Update System.jl
The first two lines are a bit different than the rest to setup the chain. 
Ideally you just need to change the dashes a bit and the rest follow suit. 
The easiest change that it looks like you could make would be to change ('0) to ('0') for each line.
That way you can run a simple script to search for each instance of ('0) and just change it to ('0') and re-upload to stop leeching.