Update The-System.jl
Should be setup to 'stream' now for Harry, Maggie, Nomi and Rin. 
There's 2 for each to make sure they're balanced on each side. 
Then below them is supposed to be how the 1's and 0's would stream out, where everything there should be tupled.
With the dots and dashes, as well as how it's tupled, it'd be intended to 'stream out' for Red, Green and Blue across the light spectrum. 

So 3 tuples, combined with binary to stream directly to someone's eye with light. That's how this should be used ideally.
This, so far, is the only way I could see 'Magic' working.
By using 'Science' to create a back end for 'Magic' that allows for 'super natural' events to occur.
Using the 10 dimensions of string theory to have a respective 1 and a 0, for 8 different people, so that the 8 of you makeup a 'byte'.

Ideally you change at least 3 things every time you 'move' or between each part of the 'wall'. 
Representing 3 Dimensions of Space.

Using the above example, Harry and Maggie's change is their Name, Gender and Class.
Where Rin and Nomi's 'change' is with their Name, Gender, and Race. Meaning they keep the same class.

Ideally most the words here would be 'leeching' off of concepts, so you wouldn't need to define most of it outright. 
Most of it would auto-define itself based off of the 'person' that's 'streaming' that 'light' into someone's eye.